Relocate 5 scaffold framework-detection helpers (Phase 3, batch 2)
Phase 3 follow-up PR for Issue #683. Moves 5 small pure-fn scaffold
framework-detection helpers (~58 LOC of bodies) out of turn.rs into
scaffold_pipeline.rs.

Moved (turn.rs -> scaffold_pipeline.rs):

  - task_requires_nextjs_scaffold (3 LOC, #[cfg(test)] only in
    production turn.rs but called by the production
    task_or_plan_requires_nextjs_scaffold below)
  - requested_scaffold_framework (12 LOC) — keyword detector that
    classifies a task string into Next / Nuxt / React or None.
  - scaffold_command_matches_framework (19 LOC) — gate that the
    LLM-issued command matches the requested framework's
    scaffolder (create-next-app / create vite + react / nuxi).
  - task_or_plan_requires_nextjs_scaffold (7 LOC) — combines the
    active-task and plan-contents checks.
  - deterministic_nextjs_scaffold_reply (17 LOC) — synthesizes the
    canned `npx create-next-app@<v>` Bash tool call when the
    deterministic scaffold path fires.

All 5 promoted to pub(super). task_requires_nextjs_scaffold stays
exposed (not #[cfg(test)] gated in the new module) since the
production task_or_plan_requires_nextjs_scaffold caller dispatches
through it. turn.rs re-imports it under #[cfg(test)] only because
the production callers all live in actor_loop_flow.rs or in the
just-moved sibling.

Adjustments:

  * scaffold_pipeline.rs: added
    - use crate::ollama::client::AssistantReply
    - use crate::ollama::xml_fallback::ToolCall

  * turn.rs: extended use super::scaffold_pipeline::{...} with the
    4 production-callable helpers; added the #[cfg(test)] gate for
    task_requires_nextjs_scaffold.

Metrics: turn.rs 31,823 -> 31,764 (-59);
scaffold_pipeline.rs 76 -> 145 (+69).
