+++ src/agent/loop_run/scaffold_pipeline.rs
@@ -36,6 +36,9 @@
 
 use std::path::PathBuf;
 
+use crate::ollama::client::AssistantReply;
+use crate::ollama::xml_fallback::ToolCall;
+
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(super) enum ScaffoldFramework {
     Next,
@@ -74,3 +77,69 @@ pub(super) const EVENT_DETERMINISTIC_FORMAT_ERROR_SMALL_EDIT: &str =
 pub(super) const EVENT_DETERMINISTIC_PYTHON_TEST_FALLBACK: &str =
     "agent.empty_workspace.deterministic_python_test_fallback";
 pub(super) const CREATE_NEXT_APP_PACKAGE_VERSION: &str = "16.2.4";
+
+pub(super) fn task_requires_nextjs_scaffold(task: &str) -> bool {
+    requested_scaffold_framework(task) == Some(ScaffoldFramework::Next)
+}
+
+pub(super) fn requested_scaffold_framework(task: &str) -> Option<ScaffoldFramework> {
+    let normalized = task.to_ascii_lowercase();
+    if normalized.contains("next.js") || normalized.contains("nextjs") {
+        Some(ScaffoldFramework::Next)
+    } else if normalized.contains("nuxt.js") || normalized.contains("nuxt") {
+        Some(ScaffoldFramework::Nuxt)
+    } else if normalized.contains("react.js") || normalized.contains("react") {
+        Some(ScaffoldFramework::React)
+    } else {
+        None
+    }
+}
+
+pub(super) fn scaffold_command_matches_framework(
+    framework: ScaffoldFramework,
+    command: &str,
+) -> bool {
+    let normalized = command.to_ascii_lowercase();
+    match framework {
+        ScaffoldFramework::Next => normalized.contains("create-next-app"),
+        ScaffoldFramework::React => {
+            (normalized.contains("create vite")
+                || normalized.contains("create-vite")
+                || normalized.contains("vite@latest")
+                || normalized.contains("vite@"))
+                && normalized.contains("react")
+        }
+        ScaffoldFramework::Nuxt => {
+            normalized.contains("nuxi")
+                || normalized.contains("create-nuxt")
+                || normalized.contains("create nuxt")
+                || normalized.contains("nuxt@")
+        }
+    }
+}
+
+pub(super) fn task_or_plan_requires_nextjs_scaffold(
+    active_task: Option<&str>,
+    plan_contents: Option<&str>,
+) -> bool {
+    active_task.is_some_and(task_requires_nextjs_scaffold)
+        || plan_contents.is_some_and(task_requires_nextjs_scaffold)
+}
+
+pub(super) fn deterministic_nextjs_scaffold_reply() -> AssistantReply {
+    let command = format!(
+        "npx --yes create-next-app@{CREATE_NEXT_APP_PACKAGE_VERSION} . --typescript --tailwind --eslint --app --no-src-dir --import-alias \"@/*\" --use-npm --yes"
+    );
+    AssistantReply {
+        content: String::new(),
+        tool_calls: vec![ToolCall {
+            id: "deterministic-nextjs-scaffold-1".to_string(),
+            name: "Bash".to_string(),
+            arguments: serde_json::json!({
+                "command": command
+            }),
+        }],
+        prompt_tokens: None,
+        completion_tokens: None,
+    }
+}
