+++ src/agent/loop_run/turn.rs
@@ -86,11 +86,15 @@ use super::safe_stop_payload::SAFE_STOP_REPORT_EVENT_MAX_BYTES;
 use super::safe_stop_payload::{build_safe_stop_payload, collect_recent_action_labels};
 #[cfg(test)]
 use super::scaffold_pipeline::PlanExplorationKey;
+#[cfg(test)]
+use super::scaffold_pipeline::task_requires_nextjs_scaffold;
 use super::scaffold_pipeline::{
     CREATE_NEXT_APP_PACKAGE_VERSION, DeterministicScaffoldSpec,
     EVENT_DETERMINISTIC_FASTAPI_SCAFFOLD, EVENT_DETERMINISTIC_FORMAT_ERROR_SMALL_EDIT,
     EVENT_DETERMINISTIC_PYTHON_CLI, EVENT_DETERMINISTIC_PYTHON_TEST_FALLBACK,
-    ScaffoldFallbackResult, ScaffoldFramework,
+    ScaffoldFallbackResult, ScaffoldFramework, deterministic_nextjs_scaffold_reply,
+    requested_scaffold_framework, scaffold_command_matches_framework,
+    task_or_plan_requires_nextjs_scaffold,
 };
 #[cfg(test)]
 use super::semantic_repair_planning::{
@@ -1630,10 +1634,6 @@ fn extract_requested_port(task: &str) -> Option<String> {
     None
 }
 
-fn task_requires_nextjs_scaffold(task: &str) -> bool {
-    requested_scaffold_framework(task) == Some(ScaffoldFramework::Next)
-}
-
 impl ScaffoldFramework {
     fn label(self) -> &'static str {
         match self {
@@ -1653,65 +1653,6 @@ impl ScaffoldFramework {
                 "Use a Nuxt scaffold, for example: npx nuxi@latest init . --packageManager npm."
             }
         }
-    }
-}
-
-fn requested_scaffold_framework(task: &str) -> Option<ScaffoldFramework> {
-    let normalized = task.to_ascii_lowercase();
-    if normalized.contains("next.js") || normalized.contains("nextjs") {
-        Some(ScaffoldFramework::Next)
-    } else if normalized.contains("nuxt.js") || normalized.contains("nuxt") {
-        Some(ScaffoldFramework::Nuxt)
-    } else if normalized.contains("react.js") || normalized.contains("react") {
-        Some(ScaffoldFramework::React)
-    } else {
-        None
-    }
-}
-
-fn scaffold_command_matches_framework(framework: ScaffoldFramework, command: &str) -> bool {
-    let normalized = command.to_ascii_lowercase();
-    match framework {
-        ScaffoldFramework::Next => normalized.contains("create-next-app"),
-        ScaffoldFramework::React => {
-            (normalized.contains("create vite")
-                || normalized.contains("create-vite")
-                || normalized.contains("vite@latest")
-                || normalized.contains("vite@"))
-                && normalized.contains("react")
-        }
-        ScaffoldFramework::Nuxt => {
-            normalized.contains("nuxi")
-                || normalized.contains("create-nuxt")
-                || normalized.contains("create nuxt")
-                || normalized.contains("nuxt@")
-        }
-    }
-}
-
-fn task_or_plan_requires_nextjs_scaffold(
-    active_task: Option<&str>,
-    plan_contents: Option<&str>,
-) -> bool {
-    active_task.is_some_and(task_requires_nextjs_scaffold)
-        || plan_contents.is_some_and(task_requires_nextjs_scaffold)
-}
-
-fn deterministic_nextjs_scaffold_reply() -> AssistantReply {
-    let command = format!(
-        "npx --yes create-next-app@{CREATE_NEXT_APP_PACKAGE_VERSION} . --typescript --tailwind --eslint --app --no-src-dir --import-alias \"@/*\" --use-npm --yes"
-    );
-    AssistantReply {
-        content: String::new(),
-        tool_calls: vec![ToolCall {
-            id: "deterministic-nextjs-scaffold-1".to_string(),
-            name: "Bash".to_string(),
-            arguments: serde_json::json!({
-                "command": command
-            }),
-        }],
-        prompt_tokens: None,
-        completion_tokens: None,
     }
 }
